Tests added for the PlaceInformation, one for the Overview and another one for the Reviews. (Also, small add to Place so that test does not break, import added)
