+++ webapp/src/domain/Place.tsx
@@ -1,3 +1,5 @@
+import * as crypto from "crypto";
+
 /**
  * Class to define a point retrieved from the SOLID dataset
  */
@@ -10,14 +12,14 @@ export default class Place {
     private _photos: File[];
     private _category: string;
 
-    constructor(name:string, latitude:number, longitude:number, description:string, photos:File[]=new Array<File>(), id:string|undefined=undefined, category: string) {
-      this._title = name;
-      this._latitude = latitude;
-      this._longitude = longitude;
-      this._description = description;
-      this._photos = photos;
-      this._uuid = (id === undefined) ? crypto.randomUUID() : id;
-      this._category = category;
+    constructor(name: string, latitude: number, longitude: number, description: string, photos: File[] = new Array<File>(), id: string | undefined = undefined, category: string) {
+        this._title = name;
+        this._latitude = latitude;
+        this._longitude = longitude;
+        this._description = description;
+        this._photos = photos;
+        this._uuid = (id === undefined) ? crypto.randomUUID() : id;
+        this._category = category;
     }
 
     get uuid(): string {
@@ -25,55 +27,55 @@ export default class Place {
     }
 
     get title(): string {
-      return this._title;
+        return this._title;
     }
-  
+
     set title(name: string) {
-      this._title = name;
+        this._title = name;
     }
-  
+
     get latitude(): number {
-      return this._latitude;
+        return this._latitude;
     }
-  
+
     set latitude(latitude: number) {
-      this._latitude = latitude;
+        this._latitude = latitude;
     }
-  
+
     get longitude(): number {
-      return this._longitude;
+        return this._longitude;
     }
-  
+
     set longitude(longitude: number) {
-      this._longitude = longitude;
+        this._longitude = longitude;
     }
-  
+
     get description(): string {
-      return this._description;
+        return this._description;
     }
-  
+
     set description(description: string) {
-      this._description = description;
+        this._description = description;
     }
-  
-    get photos(): File[]{
-      return this._photos;
+
+    get photos(): File[] {
+        return this._photos;
     }
-  
+
     set photo(photo: File) {
-      this._photos = [...this._photos, photo];
+        this._photos = [...this._photos, photo];
     }
-    
-    set photos(photos: File[]){
-      this._photos = photos;
+
+    set photos(photos: File[]) {
+        this._photos = photos;
     }
 
     set category(category: string) {
-      this._category = category;
+        this._category = category;
     }
-  
+
     get category(): string {
-      return this._category;
+        return this._category;
     }
 
 
@@ -84,4 +86,4 @@ export default class Place {
     //     // convert binary data to base64 string
     //     return new Buffer(bitmap).toString('base64');
     // }
-  }
+}
